Temporary fix for this problem:
Warning: we may have some problems with surfaces 117 and 7M in BilbaoWheel.cxx.
Indeed, surface 117 is defined as a cylinder with radius targetInnerHeightRadius + steelTungstenInnerThick, which are set to 74.5 and 0.5 respectively.
Now, if we set WheelRadius2 to 75.0, then surface 117 and 27 (7M) are coincident and, therefore, replaced with a single surface. The corresponding cell, then ends up containing something like "... 27 -27".
This does not show up as an error in the viewer, but particles get lost in there.

My temporary fix is to set WheelRadius2 to 75.01, so that the cell exists, but I think we need a better solution.